pyproject.toml
[project]
name = "memory-gym"
version = "0.1.0"
description = "MemoryGym: An end-to-end synthetic benchmark for evaluating agentic memory in LLMs"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Foundry Memory Team" }
]
keywords = ["llm", "personalization", "benchmark", "conversation", "evaluation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "openai>=1.0.0",
    "azure-identity>=1.15.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "json-repair>=0.25.0",
    "tqdm>=4.65.0",
    "timeout-decorator>=0.5.0",
    "tiktoken>=0.12.0",
    "tenacity>=9.1.2",
    "azure-ai-projects>=2.0.0b3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
memory-gym-generate = "memory_gym.data_generation:main"
memory-gym-tasks = "memory_gym.task_generator:main"
memory-gym-eval = "memory_gym.evaluation:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["memory_gym"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]  # Line too long - handled by formatter

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
